% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fourplot2.R
\name{fourplot}
\alias{fourplot}
\title{Fourplot}
\usage{
fourplot(vec)
}
\arguments{
\item{vec}{vector of data}
}
\value{
four different plots
}
\description{
Creates a fourplot of a vector
}
\details{
longer description: what the function does
}
\examples{
fourplot(data)

}
